apps/api/requirements.txt
fastapi==0.114.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
pydantic-settings==2.4.0
sqlmodel==0.0.21
psycopg2-binary==2.9.9
qdrant-client==1.9.1
meilisearch==0.31.5
neo4j==5.25.0
llama-index==0.11.0
llama-index-embeddings-huggingface==0.2.0
llama-index-vector-stores-qdrant==0.2.0
sentence-transformers==3.1.0
rapidfuzz==3.9.6
pandas==2.2.2
pyarrow==16.1.0
great_expectations==0.18.14
alembic==1.13.2
PyJWT==2.9.0
strawberry-graphql==0.227.2
